apiserver/requirements/common.txt
amqp==5.1.1
astroid==2.12.13
autopep8==2.0.0
black==22.3.0
billiard==3.6.4.0
celery==5.2.7
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
future==0.18.2
gunicorn==20.1.0
importlib-metadata==5.1.0
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.8.0
mccabe==0.7.0
pre-commit==2.20.0
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.15.6
pytz==2022.6
redis==4.3.5
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
vine==5.0.0
wrapt==1.14.1
zipp==3.11.0
openpyxl==3.0.10
pillow==9.4.0
drf-yasg==1.21.5
python-dateutil==2.8.2
django-storages==1.13.1
boto3==1.26.137